requirement.txt
streamlit
langchain
google-genai
faiss-cpu
transformers
torch
sentence-transformers
